+++ src/float/small_float.rs
@@ -277,6 +277,7 @@ impl Assign<f64> for SmallFloat {
 
 #[inline]
 fn rraw(round: Round) -> mpfr::rnd_t {
+    #[allow(deprecated)]
     match round {
         Round::Nearest => mpfr::rnd_t::RNDN,
         Round::Zero => mpfr::rnd_t::RNDZ,